UpdateCloneCreationTime now goes and turns on bitset[0] -This allows for the clone to be treated as normal items.
Component id is now a unsigned int

+++ ChronoLoop/ChronoLoop/Objects/Component.h
@@ -30,7 +30,7 @@ namespace Epoch
 		friend class BaseObject;
 
 		static unsigned short mComponentCount;
-		unsigned short mComponentId;	//unique component id
+		unsigned int mComponentId;	//unique component id
 		unsigned short mComponentNum;  //the nth number component of a base object. This is for knowing the position in the bitset
 
 	protected:
